browser and node html schema
