+++ package.json
@@ -6,7 +6,7 @@
     "prettier": "npx prettier . -w",
     "eslint": "npx eslint",
     "csslint": "npx stylelint src/**/*.css --fix",
-    "puppeteer": "node tests/puppeteer/tree-validation.js",
+    "puppeteer": "node tests/puppeteer/tree-validation.test.js",
     "static": "cp src/robots.txt dist",
     "start": "parcel && npm run static",
     "build": "parcel build --dist-dir dist --public-url /website && npm run static",
@@ -41,6 +41,9 @@
     "puppeteer": "^22.15.0",
     "stylelint": "^16.8.1",
     "stylelint-config-standard": "^36.0.1"
+  },
+  "@parcel/resolver-default": {
+    "packageExports": true
   },
   "type": "module",
   "targets": {
