+++ src/index.html
@@ -6,7 +6,6 @@
     <meta name="description" content="Home page" />
     <title>Home</title>
     <link rel="stylesheet" href="/src/styles/common.css" />
-    <script src="/src/index.js"></script>
   </head>
   <body class="body">
     <header>
