--- src/index.js
@@ -1,3 +1,0 @@
-;(async () => {
-  console.log('Hello Venus!')
-})()
