+++ src/outline/eu-accessibility-act.html
@@ -10,6 +10,7 @@
     <link rel="stylesheet" href="/src/styles/common.css" />
     <link rel="stylesheet" href="/src/styles/design-system.css" />
     <title>The European Accessibility Act</title>
+    <script src="/src/tools/html-validation.js" type="module"></script>
   </head>
   <body class="body">
     <header>
