+++ src/styles/common.css
@@ -10,8 +10,14 @@
 
 /* rules */
 
-* {
+html {
   box-sizing: border-box;
+}
+
+*,
+*::before,
+*::after {
+  box-sizing: inherit;
 }
 
 *:empty:not(input, img, hr) {
