+++ src/tools/config.js
@@ -1,0 +1,44 @@
+export const config = {
+  viewport: {
+    desktop: {
+      width: 1920,
+      height: 1080,
+      scrolled: 30000,
+    },
+    mobile: {
+      width: 1080,
+      height: 2400,
+      scrolled: 30000,
+    },
+  },
+}
+
+export const treeElements = [
+  'main',
+  'header',
+  'nav',
+  'h1',
+  'h2',
+  'h3',
+  'h4',
+  'h5',
+  'h6',
+  'aside',
+  'footer',
+  'p',
+  'ul',
+  'ol',
+  'dl',
+  'li',
+  'dt',
+  'dd',
+  'table',
+  'th',
+  'td',
+  'form',
+  'legend',
+  'figure',
+  'img',
+  'figcaption',
+  'a',
+]
