+++ src/tools/html-validation.js
@@ -1,0 +1,6 @@
+import { parseHtmlDocumentInBrowser } from '/src/tools/js/browser-validation.js'
+;(async () => {
+  const treeStructure = parseHtmlDocumentInBrowser()
+
+  console.dir(treeStructure, { depth: null })
+})()
