--- tests/config.js
@@ -1,14 +1,0 @@
-export const config = {
-  viewport: {
-    desktop: {
-      width: 1920,
-      height: 1080,
-      scrolled: 30000,
-    },
-    mobile: {
-      width: 1080,
-      height: 2400,
-      scrolled: 30000,
-    },
-  },
-}
